NmlKey class for namelist keys
This patch introduces a NmlKey class, which externally matches the
namelist key as it appears in the file, but contains a field (_key)
which holds its internal key, used when duplicate groups exist within
the namelist.

The `keys`, `items` and `__iter__` of Namelist were also modified to
return these NmlKey strings in place of the normal keys.  Views are used
in place of generic iterators and lists.

(Note: Performance gain from Views is doubtful, since we are pulling
from the `collections` module rather than faster intrinsic C
implementations, but I guess it is more "correct" than returning a list
or tuple.)

When these keys are used to access values within the namelist, they will
return the specific value for which they represent.  Externally, they
use the generic name.

Usage across namelists is not addressed here, nor any other creative
attempts to use this feature.