Check cogroups list for key existence
Checking for the existence of a key would fail if the key was a cogroup,
since cogroup keys were renamed to the internal format (`_grp_..._###`).

This was resolved by searching both the namelist dict keys and the
internal cogroup list.

This also fixed a bug with removal of elements from self._cogroups,
which was apparently untested before this point.