Pypy: Pop internal key for cogroup reorder
Some difference in the pop() implementation on PyPy requires that we
pope the internal key, rather than use the NmlKey string.  Most likely
not leaning on __delitem__ as much.

In any case, it makes more sense to pass the internal key, so it's a
reasonable change.

This patch restores PyPy support.