Cogroup initialization support
This patch resolves an issue where a Cogroup was used to re-initialize a
Namelist.

This fix is touching issues around the ambiguity of list interpetation
as either a list of derived types or a list of repeated keys.  Currently
both are represented by lists.  The latter may be represented by a
Cogroup, if it came from a previous Namelist.

This patch addresses the latter case, but not the more general case.
Some major refactoring would be needed to resolve it.