Do not pop off derived type patch assignments
This patch replaces a `pop` with `get` in a derived type patch
assignment.

While the exact behavior is not yet clear, it seems reasonable that we
would not want to prematurely pop each value, since the contents are
reused across all fields of the derived type.

There may be unexpected consequences here, but I believe this is "more
correct", and fixes some known issues.  I can only hope it does not
create any new ones.