Fix "if index"-like test with None test
Two checks related to multidimensional indices were incorrectly apply
generic `if index` tests, which return false if `index` is zero.

This raised errors in two cases:

1. The end-bound of an array a(0), which is bounded by 1, was
   incorrectly set to None rather than 1, and treated as a floating
   index.

2. The `FIndex` iterator would incorrectly terminate higher ranks if the
   current higher rank index was zero.

These errors occur because we use `None` when the index bounds are
unspecified, and are implicitly determined when values are provided.

Both of these tests have been replaced with `if None` tests, which
appear to have resolved these errors.