Add update method for cogroups
Cogroups, although subclasses of lists, are designed to behave as dicts
in many situations.  One absence is the support of certain intrinsic
methods.

The `patch` method of `Namelist` relied on `update`, but this was not
supported by Cogroups and would raise an error if one was returned.
This patch adds support to Cogroups.

It does not address which elements of the Cogroup should receive the
patch.  Currently, all elements in the cogroup are patched.  This is
sufficient to avoid errors, but should be addressed in the future.