DFA-based tokenization
This patch replaces the hand-scanner with a more formal DFA-based
scanner.  Lexemes are constructed per-character and state is tracked
with a dictionary (`M`) rather than complex if-statements and other
awkward lookahead-like constructs.

The major change here is that whitespace is now formally tokenized
alongside the semantic content.  Endlines (`\n`) are no longer formally
used for much of anything and are globbed into whitespace other other
content.  Comments are also globbed into whitespace.  These are now
gathered as "blank" tokens.

This is not a tokenizer.  Not yet at least.  It is still only a scanner.
We do "learn" something about the lexemes we are gathering, but we don't
bother to tag them as tokens.  We still re-do this in the (hand) parser.

A bug related to multiline strings has been fixed.  Newline characters
are now correctly saved as part of the string.  The test was not
catching this error and has also been updated.

NOTE: This could be a very epochal patch.  Although every test has
passed, there is still enormous room for error, and I do not believe we
catch every possible case.  This is going to need some time before it
can be moved into the main branch.