setup.py: Move content to setup.cfg
This moves most of the content in setup.py to setup.cfg.

Only the project name and the version parser are left in setup.py.

Not really sure about the whole classifiers thing but we'll come back to
it.

Added a make rule to test the package (unsure about the name)