Scanner as generator
This patch replaces the scanner with yield-based generator.

Functionally it is largely identical, but it now returns strings rather
than an entire list.

This is ok for our purposes since we were just turning it back into an
iterator anyway.

Microbenchmarks suggest that this is slightly faster; although there is
greater overhead from enumerate() and yield, it seems to be compensated
by the removal of append() and perhaps simpler comment length tracking.

Long-term, this is should eventually yield true speedups in the parser,
when it gains the ability to work with tokens rather than reinterpreting
its type based on the current state (and occasional lookahead).  But
we're not there yet.