Backwards compatible scanner
This patch "de-versions" the scanner to work on older versions of
Python. We did not mean to drop Python 2.7 suddenly, so this removes the
following constructs from the scanner:

* No `d1 | d2` merges of dicts.

* No use of "walrus" operator, `|=`

Perhaps a lot of work for nothing, since we plan to drop Py2.7 very
soon, but at least it will be on purpose!