Remove redundant commas of tail null values
Certain compilers require null value assignments to end with commas.
The `Namelist` writer contains a special rule to append this comma.

If `end_comma` is enabled, which always finalizes values with commas,
then this would append a second comma, which would transform the record
into two null values.

This patch prevents the redundant comma if end_commas are enabled.

Thanks to Paul Leopardi (@penguian) for reporting.